Update to LTS-8.17.
The motivation for this change is that purify doesn't work behind a proxy. That's because http-client-tls doesn't respect `http_proxy` and `https_proxy` environment variables until v0.3.4.

https://www.stackage.org/lts-8.17/package/http-client-tls-0.3.5#changes

Updating to http-client-tls >= 0.3.4 fixes the problem. The simplest way to do that is just update to >= LTS-8.3. (In this commit I've updated to the latest LTS available though.)